+++ examples/hello_asyncio.py
@@ -29,8 +29,8 @@ import sys
 # to load the module from there first.
 basedir = os.path.abspath(os.path.join(os.path.dirname(sys.argv[0]), '..'))
 if (os.path.exists(os.path.join(basedir, 'setup.py')) and
-    os.path.exists(os.path.join(basedir, 'src', 'pyfuse3', 'pyfuse3.pyx'))):
-    sys.path.insert(0, os.path.join(basedir, 'src', 'pyfuse3'))
+    os.path.exists(os.path.join(basedir, 'src', 'pyfuse3', '__init__.pyx'))):
+    sys.path.insert(0, os.path.join(basedir, 'src'))
 
 from argparse import ArgumentParser
 import asyncio
@@ -38,7 +38,7 @@ import stat
 import logging
 import errno
 import pyfuse3
-import pyfuse3_asyncio
+import pyfuse3.asyncio
 
 try:
     import faulthandler
@@ -48,7 +48,7 @@ else:
     faulthandler.enable()
 
 log = logging.getLogger(__name__)
-pyfuse3_asyncio.enable()
+pyfuse3.asyncio.enable()
 
 class TestFs(pyfuse3.Operations):
     def __init__(self):